# backend/openapi.json
{
  "components": {
    "responses": {
      "Unauthorized": {
        "description": "Authentication credentials are missing, malformed, expired, or invalid for this operation."
      }
    },
    "schemas": {
...
      "DataBaseTimestampReturn": {
        "properties": {
          "abstractData": {
            "type": "object"
          },
          "assetId": {
            "description": "Path-primary asset ID.",
            "type": "string"
          },
          "coverHash": {
            "description": "For albums: the cover image's content hash (used for compressed\nthumbnail URL construction and token validation). `None` for media\nitems or albums without a cover.",
            "type": [
              "string",
              "null"
            ]
          },
          "timestamp": {
            "format": "int64",
            "type": "integer"
          },
          "token": {
            "type": "string"
          }
        },
        "required": [
          "abstractData",
          "timestamp",
          "token",
          "assetId"
        ],
        "type": "object"
      },
...
      "DisplayElement": {
        "properties": {
          "displayHeight": {
            "format": "int32",
            "minimum": 0,
            "type": "integer"
          },
          "displayWidth": {
            "format": "int32",
            "minimum": 0,
            "type": "integer"
          }
        },
        "required": [
          "displayWidth",
          "displayHeight"
        ],
        "type": "object"
      },
...
      "FsCompletion": {
        "properties": {
          "children": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "is_default": {
            "type": "boolean"
          },
          "roots": {
            "items": {
              "type": "string"
            },
            "type": "array"
          }
        },
        "required": [
          "roots",
          "children",
          "is_default"
        ],
        "type": "object"
      },
...
      "Row": {
        "properties": {
          "displayElements": {
            "items": {
              "$ref": "#/components/schemas/DisplayElement"
            },
            "type": "array"
          },
          "end": {
            "minimum": 0,
            "type": "integer"
          },
          "rowIndex": {
            "minimum": 0,
            "type": "integer"
          },
          "start": {
            "minimum": 0,
            "type": "integer"
          }
        },
        "required": [
          "start",
          "end",
          "displayElements",
          "rowIndex"
        ],
        "type": "object"
      },
      "ScrollBarData": {
        "properties": {
          "index": {
            "minimum": 0,
            "type": "integer"
          },
          "month": {
            "minimum": 0,
            "type": "integer"
          },
          "year": {
            "minimum": 0,
            "type": "integer"
          }
        },
        "required": [
          "year",
          "month",
          "index"
        ],
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "",
    "license": {
      "identifier": "MIT",
      "name": "MIT"
    },
    "title": "picasu",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/get/get-data": {
      "get": {
        "description": "Phase 14 lean read path: media rows are built from the snapshot's\n`ReducedData` plus the lean `ASSET_BY_ID` record — no per-row\n`METADATA_TABLE` (payload) read, and no tags/EXIF/description on the\npayload (those are served by `GET /get/metadata/{assetId}`).\nAlbum rows still read `METADATA_TABLE` because tiles need their stored\ntitle/cover/counts, composed with the album's `AssetRecord`.",
        "operationId": "get_data",
        "parameters": [
          {
            "in": "query",
            "name": "end",
            "required": true,
            "schema": {
              "minimum": 0,
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "start",
            "required": true,
            "schema": {
              "minimum": 0,
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "timestamp",
            "required": true,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/DataBaseTimestampReturn"
                  },
                  "type": "array"
                }
              }
            },
            "description": "Data by timestamp range"
          },
          "400": {
            "description": "Invalid input"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        },
        "summary": "Serve one page of timeline/list rows for a snapshot timestamp.",
        "tags": [
          "timeline"
        ]
      }
    },
...
    "/get/get-rows": {
      "get": {
        "operationId": "get_rows",
        "parameters": [
          {
            "in": "query",
            "name": "index",
            "required": true,
            "schema": {
              "minimum": 0,
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "timestamp",
            "required": true,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Row"
                }
              }
            },
            "description": "Row data"
          },
          "400": {
            "description": "Invalid input"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        },
        "summary": "Serve one row of a snapshot by its index.",
        "tags": [
          "timeline"
        ]
      }
    },
    "/get/get-scroll-bar": {
      "get": {
        "operationId": "get_scroll_bar",
        "parameters": [
          {
            "in": "query",
            "name": "timestamp",
            "required": true,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ScrollBarData"
                  },
                  "type": "array"
                }
              }
            },
            "description": "Scroll bar data"
          },
          "400": {
            "description": "Invalid input"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        },
        "summary": "Serve the scroll bar positions of a snapshot.",
        "tags": [
          "timeline"
        ]
      }
    },
...
    "/get/metadata/{asset_id}": {
      "get": {
        "description": "This is the detail-side counterpart of `get-data`: list rows only carry\nlean identity fields (tags / EXIF / description / rating are stripped in\nPhase 14), so the sidebar, detail view, and edit prefill fetch the full\n`AbstractData` view here on demand. The wire shape is unchanged:\nidentity fields come from the record, metadata fields from the payload.\n\nAuth and share parity follow `get-data`: a `GuardTimestamp` bearer token\n(prefetch token) is required, and when the token resolves to a share with\n`show_metadata: false` the metadata fields are cleared before responding so\na share that hides metadata cannot leak it through this route.",
        "operationId": "get_metadata",
        "parameters": [
          {
            "in": "path",
            "name": "asset_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "timestamp",
            "required": true,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Full metadata record for the asset"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "description": "Unknown asset_id"
          }
        },
        "summary": "Full metadata detail for a single asset, composed from its identity and its stored metadata payload.",
        "tags": [
          "assets"
        ]
      }
    },
    "/get/path-completion": {
      "get": {
        "operationId": "get_fs_completion",
        "parameters": [
          {
            "in": "query",
            "name": "path",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FsCompletion"
                }
              }
            },
            "description": "Filesystem path completion"
          },
          "400": {
            "description": "Invalid input"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        },
        "summary": "Complete a filesystem path for the path-completion field.",
        "tags": [
          "config"
        ]
      }
    },
...
  }
}