# backend/src/router/get/get_data.rs
// src/router/get/get_data.rs

use crate::model::asset::{AssetKind, AssetRecord};
use crate::model::metadata_record::compose_abstract_data;
use crate::model::response::DataBaseTimestampReturn;
use crate::model::response::{Row, ScrollBarData};
use crate::openapi_components::Unauthorized;
use crate::process::resolve_show_download_and_metadata;
use crate::process::transitor::{cover_content_hash_from_data, lean_media_abstract_data};
use crate::storage::cache::{SnapshotReadError, TREE_SNAPSHOT};
use crate::storage::db::{ASSET_BY_ID, TREE, open_metadata_table, open_tree_snapshot_table};

use crate::error::{AppError, ErrorKind, ResultExt};
use crate::router::auth::GuardTimestamp;
use crate::router::{AppResult, GuardResult};
use anyhow::Result;
use log::info;
use rayon::iter::{IntoParallelIterator, ParallelIterator};
use redb::ReadableDatabase;
use rocket::serde::json::Json;
use std::time::Instant;

/// Maps a tree-snapshot read failure onto the HTTP error model.
///
/// A snapshot id that is absent from both the in-memory map and the on-disk
/// store was never minted or has been dropped by the ~1h expire check, which
/// makes it client input: it maps to `ErrorKind::InvalidInput` (400), matching
/// the `(status = 400, description = "Invalid input")` both operations already
/// publish. Everything the snapshot store reports as `Storage` is a genuine
/// server-side fault and stays `ErrorKind::Database` (500).
fn map_snapshot_read_error(err: SnapshotReadError) -> AppError {
    match err {
        SnapshotReadError::NotFound { timestamp } => AppError::new(
            ErrorKind::InvalidInput,
            format!("Unknown snapshot id: {timestamp}"),
        ),
        SnapshotReadError::Storage(source) => AppError::from_err(ErrorKind::Database, source),
    }
}

/// Serve one page of timeline/list rows for a snapshot timestamp.
///
/// Phase 14 lean read path: media rows are built from the snapshot's
/// `ReducedData` plus the lean `ASSET_BY_ID` record — no per-row
/// `METADATA_TABLE` (payload) read, and no tags/EXIF/description on the
/// payload (those are served by `GET /get/metadata/{assetId}`).
/// Album rows still read `METADATA_TABLE` because tiles need their stored
/// title/cover/counts, composed with the album's `AssetRecord`.
#[utoipa::path(
        tag = "timeline",
        responses(
            (status = 200, description = "Data by timestamp range", body = Vec<DataBaseTimestampReturn>),
            (status = 400, description = "Invalid input"),
            (status = 401, response = Unauthorized),
        )
    )
]
#[get("/get/get-data?<timestamp>&<start>&<end>")]
pub async fn get_data(
    guard_timestamp: GuardResult<GuardTimestamp>,
    timestamp: i64,
    start: usize,
    mut end: usize,
) -> AppResult<Json<Vec<DataBaseTimestampReturn>>> {
    let guard_timestamp = guard_timestamp?;
    tokio::task::spawn_blocking(move || {
        let start_time = Instant::now();

        let resolved_share_opt = guard_timestamp.claims.resolved_share_opt;
        let (show_download, show_metadata) = resolve_show_download_and_metadata(resolved_share_opt);

        let metadata_table = open_metadata_table();
        let tree_snapshot = open_tree_snapshot_table(timestamp)
            .or_raise(|| (ErrorKind::Database, "Failed to open tree snapshot table"))?;

        // Lean identity records for media rows (path, kind, times, album).
        let asset_txn = TREE
            .in_disk
            .begin_read()
            .or_raise(|| (ErrorKind::Database, "Failed to begin read transaction"))?;
        let asset_by_id = asset_txn
            .open_table(ASSET_BY_ID)
            .or_raise(|| (ErrorKind::Database, "Failed to open ASSET_BY_ID"))?;

        end = end.min(tree_snapshot.len().map_err(map_snapshot_read_error)?);

        if start >= end {
            return Ok(Json(vec![]));
        }

        let database_timestamp_return_list: Result<Vec<_>, AppError> = (start..end)
            .into_par_iter()
            .map(|index| {
                let reduced = tree_snapshot.get_reduced(index).or_raise(|| {
                    (
                        ErrorKind::Database,
                        format!("Failed to read snapshot entry for index {index}"),
                    )
                })?;
                let asset_id = reduced.asset_id;

                let record_json = asset_by_id
                    .get(&*asset_id)
                    .or_raise(|| {
                        (
                            ErrorKind::Database,
                            format!("Failed to read ASSET_BY_ID for asset_id {asset_id}"),
                        )
                    })?
                    .ok_or_else(|| {
                        AppError::new(
                            ErrorKind::Database,
                            format!("No ASSET_BY_ID record for asset_id {asset_id}"),
                        )
                    })?;
                let record: AssetRecord =
                    serde_json::from_str(record_json.value()).or_raise(|| {
                        (
                            ErrorKind::Database,
                            format!("Failed to parse AssetRecord for asset_id {asset_id}"),
                        )
                    })?;

                let (abstract_data, cover_content_hash) = match record.kind {
                    // Albums keep their stored payload: tiles need
                    // title/cover/counts; identity (dir_path, id, trash)
                    // is composed from the AssetRecord.
                    AssetKind::Album => {
                        let payload = metadata_table
                            .get(&*asset_id)
                            .or_raise(|| {
                                (
                                    ErrorKind::Database,
                                    format!("Failed to read album payload for asset_id {asset_id}"),
                                )
                            })?
                            .ok_or_else(|| {
                                AppError::new(
                                    ErrorKind::Database,
                                    format!("Failed to retrieve album for asset_id {asset_id}"),
                                )
                            })?
                            .value();
                        let abstract_data = compose_abstract_data(&record, Some(&payload));
                        let cover = cover_content_hash_from_data(&abstract_data);
                        (abstract_data, cover)
                    }
                    // Media rows: lean construction, no metadata read.
                    AssetKind::Image | AssetKind::Video => {
                        (lean_media_abstract_data(&record, &reduced), None)
                    }
                };

                let mut database_timestamp_return =
                    crate::process::transitor::abstract_data_to_timestamp_return(
                        abstract_data,
                        timestamp,
                        show_download,
                        show_metadata,
                        asset_id,
                        cover_content_hash,
                    );
                // Row timestamps equal the tree snapshot's date (computed
                // from the full in-memory record when the snapshot was taken),
                // so lean rows keep the EXIF-derived sort date.
                database_timestamp_return.timestamp = reduced.date;
                Ok(database_timestamp_return)
            })
            .collect();

        let duration = format!("{:?}", start_time.elapsed());
        info!(duration = &*duration; "Get data: {start} ~ {end}");
        Ok(Json(database_timestamp_return_list?))
    })
    .await
    .or_raise(|| (ErrorKind::Internal, "Failed to join blocking task"))?
}

/// Serve one row of a snapshot by its index.
#[utoipa::path(
        tag = "timeline",
        responses(
            (status = 200, description = "Row data", body = Row),
            (status = 400, description = "Invalid input"),
            (status = 401, response = Unauthorized),
        )
    )
]
#[get("/get/get-rows?<index>&<timestamp>")]
pub async fn get_rows(
    auth: GuardResult<GuardTimestamp>,
    index: usize,
    timestamp: i64,
) -> AppResult<Json<Row>> {
    let _ = auth?;
    tokio::task::spawn_blocking(move || {
        let start_time = Instant::now();
        let filtered_rows = TREE_SNAPSHOT
            .read_row(index, timestamp)
            .map_err(map_snapshot_read_error)?;
        let duration = format!("{:?}", start_time.elapsed());
        info!(duration = &*duration; "Read rows: index = {index}");
        Ok(Json(filtered_rows))
    })
    .await
    .or_raise(|| (ErrorKind::Internal, "Failed to join blocking task"))?
}

/// Serve the scroll bar positions of a snapshot.
#[utoipa::path(
        tag = "timeline",
        responses(
            (status = 200, description = "Scroll bar data", body = Vec<ScrollBarData>),
            (status = 400, description = "Invalid input"),
            (status = 401, response = Unauthorized),
        )
    )
]
#[get("/get/get-scroll-bar?<timestamp>")]
pub fn get_scroll_bar(
    auth: GuardResult<GuardTimestamp>,
    timestamp: i64,
) -> AppResult<Json<Vec<ScrollBarData>>> {
    let _ = auth?;
    let scrollbar_data = TREE_SNAPSHOT
        .read_scrollbar(timestamp)
        .map_err(map_snapshot_read_error)?;
    Ok(Json(scrollbar_data))
}